Мерж из 0.2.1
git-svn-id: https://horse.aplana.com/svn/SBRFACCTAX/trunk/taxaccounting@1895 34aa7802-8fb0-764d-beb8-9b5b9d595954

+++ gwtapp/src/main/java/com/aplana/sbrf/taxaccounting/web/module/declarationdata/client/DeclarationDataPresenter.java
@@ -17,6 +17,7 @@ import com.aplana.sbrf.taxaccounting.web.module.declarationdata.shared.GetDeclar
 import com.aplana.sbrf.taxaccounting.web.module.declarationdata.shared.RefreshDeclarationDataAction;
 import com.aplana.sbrf.taxaccounting.web.module.declarationdata.shared.RefreshDeclarationDataResult;
 import com.aplana.sbrf.taxaccounting.web.module.declarationlist.client.DeclarationListNameTokens;
+import com.aplana.sbrf.taxaccounting.web.widget.pdfviewer.shared.Pdf;
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.user.client.Window;
 import com.google.inject.Inject;
@@ -44,18 +45,19 @@ public class DeclarationDataPresenter extends Presenter<DeclarationDataPresenter
 	}
 
 	public interface MyView extends View, HasUiHandlers<DeclarationDataUiHandlers> {
-		void setShowAccept(boolean show);
-		void setShowReject(boolean show);
-		void setShowRefresh(boolean show);
-		void setShowDownloadXml(boolean show);
-		void setShowDelete(boolean show);
+		void showAccept(boolean show);
+		void showReject(boolean show);
+		void showRefresh(boolean show);
+		void showDownloadXml(boolean show);
+		void showDelete(boolean show);
 		void setType(String type);
 		void setTitle(String title);
 		void setDepartment(String department);
 		void setReportPeriod(String reportPeriod);
 		void setBackButton(String link);
-		void setPdfFile(String fileUrl);
-		void setVisiblePdfFile(boolean visible);
+		void setPdf(Pdf pdf);
+		void setDocDate(String date);
+		void showPdfFile(boolean show);
 		void clearPdfFile();
 	}
 
@@ -83,11 +85,11 @@ public class DeclarationDataPresenter extends Presenter<DeclarationDataPresenter
 	@Override
 	public void prepareFromRequest(PlaceRequest request) {
 		super.prepareFromRequest(request);
-		final long declarationId = Integer.valueOf(request.getParameter(DeclarationDataTokens.declarationId, "0"));
-		if (declarationId != 0) {
-			this.declarationId = declarationId;
+		final long id = Integer.valueOf(request.getParameter(DeclarationDataTokens.declarationId, "0"));
+		if (id != 0) {
+			this.declarationId = id;
 			GetDeclarationDataAction action = new GetDeclarationDataAction();
-			action.setId(declarationId);
+			action.setId(id);
 			dispatcher.execute(action, CallbackUtils
 					.defaultCallback(new AbstractCallback<GetDeclarationDataResult>() {
 						@Override
@@ -95,18 +97,19 @@ public class DeclarationDataPresenter extends Presenter<DeclarationDataPresenter
 							taxName = result.getTaxType().name();
 							getView().setType("Декларация");
 							getView().setReportPeriod(result.getReportPeriod());
+							getView().setDocDate(result.getDocDate());
 							getView().setDepartment(result.getDepartment());
 							getView().setBackButton("#" + DeclarationListNameTokens.DECLARATION_LIST + ";nType="
 									+ String.valueOf(result.getTaxType()));
 							getView().setTitle(result.getTaxType().getName() + " / " + result.getDeclarationType());
 							updateTitle(result.getDeclarationType());
-							loadPdfFile();
 
-							getView().setShowAccept(result.isCanAccept());
-							getView().setShowReject(result.isCanReject());
-							getView().setShowDownloadXml(result.isCanDownload());
-							getView().setShowDelete(result.isCanDelete());
-							getView().setShowRefresh(result.isCanDelete());
+							getView().showAccept(result.isCanAccept());
+							getView().showReject(result.isCanReject());
+							getView().showDownloadXml(result.isCanDownload());
+							getView().showDelete(result.isCanDelete());
+							getView().showRefresh(result.isCanDelete());
+							getView().setPdf(result.getPdf());
 						}
 					}).addCallback(
 							new ManualRevealCallback<DeclarationDataPresenter>(
@@ -125,10 +128,11 @@ public class DeclarationDataPresenter extends Presenter<DeclarationDataPresenter
 	}
 
 	@Override
-	public void refreshDeclaration(){
+	public void refreshDeclaration(String docDate){
 		LogCleanEvent.fire(this);
 		RefreshDeclarationDataAction action = new RefreshDeclarationDataAction();
 		action.setDeclarationId(declarationId);
+		action.setDocDate(docDate);
 		dispatcher.execute(action, CallbackUtils
 				.defaultCallback(new AbstractCallback<RefreshDeclarationDataResult>() {
 					@Override
@@ -162,24 +166,29 @@ public class DeclarationDataPresenter extends Presenter<DeclarationDataPresenter
 
 	@Override
 	public void delete() {
-		LogCleanEvent.fire(this);
-		DeleteDeclarationDataAction action = new DeleteDeclarationDataAction();
-		action.setDeclarationId(declarationId);
-		dispatcher.execute(action, CallbackUtils
-				.defaultCallback(new AbstractCallback<DeleteDeclarationDataResult>() {
-					@Override
-					public void onSuccess(DeleteDeclarationDataResult result) {
-						MessageEvent.fire(DeclarationDataPresenter.this, "Декларация удалена");
-						placeManager
-								.revealPlace(new PlaceRequest(DeclarationListNameTokens.DECLARATION_LIST).with("nType", taxName));
-					}
-				}));
+		getView().showPdfFile(false);
+		if (Window.confirm("Вы уверены, что хотите удалить декларацию?")) {
+			LogCleanEvent.fire(this);
+			DeleteDeclarationDataAction action = new DeleteDeclarationDataAction();
+			action.setDeclarationId(declarationId);
+			dispatcher.execute(action, CallbackUtils
+					.defaultCallback(new AbstractCallback<DeleteDeclarationDataResult>() {
+						@Override
+						public void onSuccess(DeleteDeclarationDataResult result) {
+							MessageEvent.fire(DeclarationDataPresenter.this, "Декларация удалена");
+							placeManager
+									.revealPlace(new PlaceRequest(DeclarationListNameTokens.DECLARATION_LIST).with("nType", taxName));
+						}
+					}));
+		} else {
+			getView().showPdfFile(true);
+		}
 	}
 
 	@Override
 	@ProxyEvent
 	public void onChangeVisibility(DialogBoxChangeVisibilityEvent event) {
-		getView().setVisiblePdfFile(!event.isVisible());
+		getView().showPdfFile(!event.isVisible());
 	}
 
 	@Override
@@ -189,16 +198,12 @@ public class DeclarationDataPresenter extends Presenter<DeclarationDataPresenter
 
 	@Override
 	public void downloadExcel() {
-		Window.open(GWT.getHostPageBaseURL() + "download/downloadExcel/" + declarationId, null, null);
+		Window.open(GWT.getHostPageBaseURL() + "download/declarationData/xlsx/" + declarationId, null, null);
 	}
 
 	@Override
 	public void downloadXml() {
-		Window.open(GWT.getHostPageBaseURL() + "download/downloadXml/" + declarationId, null, null);
-	}
-
-	private void loadPdfFile() {
-		getView().setPdfFile(GWT.getHostPageBaseURL() + "download/downloadPDF/" + declarationId);
+		Window.open(GWT.getHostPageBaseURL() + "download/declarationData/xml/" + declarationId, null, null);
 	}
 
 	private void revealPlaceRequest() {
